1. HOST_NAME and POD_name is added as environment variables in deployment yaml file for all services 2. Logging framework implementation changed
